Organized packages in order to work with Travis <configuration
