+++ src/main/board/Cell.java
@@ -1,4 +1,4 @@
-package board;
+package main.board;
 
 /**
  * Creates a cell object.
