+++ src/main/board/Direction.java
@@ -1,4 +1,4 @@
-package board;
+package main.board;
 
 import java.util.Arrays;
 import java.util.Collections;
