+++ src/main/board/PowerGem.java
@@ -1,4 +1,4 @@
-package board;
+package main.board;
 
 import javafx.scene.image.Image;
