+++ src/main/game/Observer.java
@@ -1,4 +1,4 @@
-package game;
+package main.game;
 
 /**
  * Observer class for observing changes to an object.
