+++ src/main/game/Player.java
@@ -1,4 +1,4 @@
-package game;
+package main.game;
 
 import java.util.ArrayList;
 import java.util.List;
