+++ src/test/LoggerTest.java
@@ -1,4 +1,4 @@
-package tests;
+package test;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -15,7 +15,7 @@ import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import game.Logger;
+import main.game.Logger;
 
 /**
  *
